requirements.txt
beautifulsoup4==4.10.0
Pillow==9.0.1
requests==2.27.1
